Add create deployment dialog + cleanup
