--- ui/src/pages/deployment-details/deployment-details-form/types.ts
@@ -1,8 +1,0 @@
-export interface DeploymentDetailsFieldValues {
-  name: string
-  device: string
-  latitude: number
-  longitude: number
-  path: string
-  site: string
-}
