+++ ui/src/pages/deployment-details/styles.module.scss
