+++ ui/src/utils/language.ts
@@ -28,6 +28,7 @@ export enum STRING {
   DETAILS_LABEL_LONGITUDE,
   DETAILS_LABEL_LIGHT_SOURCE,
   DETAILS_LABEL_NAME,
+  DETAILS_LABEL_NEW_DEPLOYMENT,
   DETAILS_LABEL_OCCURRENCES,
   DETAILS_LABEL_PATH,
   DETAILS_LABEL_SESSION,
@@ -121,6 +122,7 @@ const ENGLISH_STRINGS: { [key in STRING]: string } = {
   [STRING.DETAILS_LABEL_LONGITUDE]: 'Longitude',
   [STRING.DETAILS_LABEL_LIGHT_SOURCE]: 'Light source',
   [STRING.DETAILS_LABEL_NAME]: 'Name',
+  [STRING.DETAILS_LABEL_NEW_DEPLOYMENT]: 'New deployment',
   [STRING.DETAILS_LABEL_OCCURRENCES]: 'Occurrences',
   [STRING.DETAILS_LABEL_PATH]: 'Path',
   [STRING.DETAILS_LABEL_SESSION]: 'Session',
